refactor(admin): extract AgentRunDetailPage's run-timeline derivation into a tested pure module (#1778)
AgentRunDetailPage recomputed its ~500-line run timeline (event-shape
normalization, orchestrator-node detection, task status, unifiedTasks
assembly) inline on every render, with zero test coverage despite being
the most fragile and frequently-patched part of the page.

Move that logic into deriveRunTimeline() in
client/src/features/admin/utils/runTimeline.js, following the tokenStats.js
pattern (plain, React-free, unit-tested), and memoize the call in the
component with useMemo. Add runTimeline.test.js covering the live-SSE vs
persisted-API event shapes, resume/retry status handling, _taskResults
backfill, planner-status independence, recovered-task ordering, and
orchestrator-node detection heuristics documented in the original inline
comments.

Also drops the unused subWorkflowState computation (confirmed dead: it was
assigned but never read anywhere in the file).

No behavior change intended - this is a pure refactor.
